Remove redundant report and address comments
