+++ cmd/util/cmd/execution-state-extract/execution_state_extract.go
@@ -93,10 +93,6 @@ func extractExecutionState(
 				Chain:                   chain,
 				PreviousStateCommitment: targetHash,
 			},
-			&reporters.EpochCounterReporter{
-				Log:   log,
-				Chain: chain,
-			},
 			&reporters.AccountReporter{
 				Log:   log,
 				Chain: chain,
