+++ cmd/util/ledger/reporters/export_reporter.go
@@ -2,10 +2,17 @@ package reporters
 
 import (
 	"encoding/json"
+	"fmt"
 	"io/ioutil"
 
 	"github.com/rs/zerolog"
 
+	"github.com/onflow/flow-core-contracts/lib/go/templates"
+
+	"github.com/onflow/flow-go/cmd/util/ledger/migrations"
+	"github.com/onflow/flow-go/fvm"
+	"github.com/onflow/flow-go/fvm/programs"
+	"github.com/onflow/flow-go/fvm/systemcontracts"
 	"github.com/onflow/flow-go/ledger"
 	"github.com/onflow/flow-go/model/flow"
 )
@@ -30,32 +37,62 @@ func (e *ExportReporter) Name() string {
 func (e *ExportReporter) Report(payload []ledger.Payload, o ledger.ExportOutputs) error {
 	script, _, err := ExecuteCurrentEpochScript(e.Chain, payload)
 
-	if err == nil && script.Err == nil && script.Value != nil {
-		epochCounter := script.Value.ToGoValue().(uint64)
+	if err != nil {
 		e.Log.
-			Info().
-			Uint64("epochCounter", epochCounter).
-			Msg("Fetched epoch counter from the FlowEpoch smart contract")
-
-		report := ExportReport{
-			CurrentStateCommitment:  o.CurrentStateCommitement,
-			EpochCounter:            script.Value.ToGoValue().(uint64),
-			PreviousStateCommitment: e.PreviousStateCommitment,
-		}
-		file, _ := json.MarshalIndent(report, "", " ")
-		e.Log.
-			Info().
-			Str("ExportReport", string(file)).
-			Msg("Ledger Export has finished")
-		_ = ioutil.WriteFile("export_report.json", file, 0644)
+			Error().
+			Err(err).
+			Msg("error running GetCurrentEpochCounter script")
+		// Safely exit and move on to next reporter
 		return nil
-	} else {
+	}
+
+	if script.Err != nil && script.Value == nil {
 		e.Log.
 			Error().
 			Err(script.Err).
 			Msg("Failed to get epoch counter")
+
+		// Safely exit and move on to next reporter
 		return nil
 	}
+
+	epochCounter := script.Value.ToGoValue().(uint64)
+	e.Log.
+		Info().
+		Uint64("epochCounter", epochCounter).
+		Msg("Fetched epoch counter from the FlowEpoch smart contract")
+
+	report := ExportReport{
+		CurrentStateCommitment:  o.CurrentStateCommitement,
+		EpochCounter:            script.Value.ToGoValue().(uint64),
+		PreviousStateCommitment: e.PreviousStateCommitment,
+	}
+	file, _ := json.MarshalIndent(report, "", " ")
+	e.Log.
+		Info().
+		Str("ExportReport", string(file)).
+		Msg("Ledger Export has finished")
+	_ = ioutil.WriteFile("export_report.json", file, 0644)
+	return nil
+}
+
+// Executes script to get current epoch from chain
+func ExecuteCurrentEpochScript(c flow.Chain, payload []ledger.Payload) (*fvm.ScriptProcedure, flow.Address, error) {
+	l := migrations.NewView(payload)
+	prog := programs.NewEmptyPrograms()
+	vm := fvm.NewVirtualMachine(fvm.NewInterpreterRuntime())
+	ctx := fvm.NewContext(zerolog.Nop(), fvm.WithChain(c))
+
+	sc, err := systemcontracts.SystemContractsForChain(c.ChainID())
+	if err != nil {
+		return nil, flow.Address{}, fmt.Errorf("error getting SystemContracts for chain %s: %w", c.String(), err)
+	}
+	address := sc.Epoch.Address
+	scriptCode := templates.GenerateGetCurrentEpochCounterScript(templates.Environment{
+		EpochAddress: address.Hex(),
+	})
+	script := fvm.Script(scriptCode)
+	return script, address, vm.Run(ctx, script, l, prog)
 }
 
 var _ ledger.Reporter = &ExportReporter{}
